# public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AccuFinance API",
    "version": "1.0.0",
    "description": "REST API for the AccuFinance Manufacturing ERP system — designed for Make-to-Order (MTO) garment production. All financial transactions use double-entry accounting with live balance derivation from journal entries. Import into Postman, Insomnia, or any OpenAPI-compatible tool.",
    "contact": {
      "name": "AccuFinance Support"
    },
    "license": {
      "name": "Proprietary"
    }
  },
  "servers": [
    {
      "url": "",
      "description": "Current server (auto-detected from where this spec is hosted)"
    }
  ],
  "security": [
    { "bearerAuth": [] },
    { "cookieAuth": [] }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "API key set via Vercel env var (API_SECRET or NEXTAUTH_SECRET). Use for Postman, curl, scripts."
      },
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "next-auth.session-token",
        "description": "NextAuth session cookie from browser login."
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean", "example": false },
          "error": { "type": "string", "example": "Authentication required" }
        }
      },
...
      "Pagination": {
        "type": "object",
        "properties": {
          "data": { "type": "array", "items": {} },
          "nextCursor": { "type": "string", "nullable": true },
          "hasMore": { "type": "boolean" }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/dashboard": {
      "get": {
        "tags": ["Dashboard"],
        "summary": "Get dashboard KPIs",
        "description": "Returns aggregated KPI data: total revenue, orders, work order status, top customers, inventory alerts, and monthly revenue trend.",
        "responses": {
          "200": {
            "description": "Dashboard data",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/DashboardResponse" } } }
          },
          "401": { "description": "Unauthorized", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
    "/api/customers": {
      "get": {
        "tags": ["Customers"],
        "summary": "List customers",
        "description": "Cursor-based pagination.",
        "parameters": [
          { "name": "limit", "in": "query", "schema": { "type": "number", "default": 50 } },
          { "name": "cursor", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Customer list", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Pagination" } } } }
        }
      },
      "post": {
        "tags": ["Customers"],
        "summary": "Create customer",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name"],
                "properties": {
                  "name": { "type": "string" },
                  "email": { "type": "string" },
                  "phone": { "type": "string" },
                  "address": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Created", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Customer" } } } }
        }
      },
      "put": {
        "tags": ["Customers"],
        "summary": "Update customer",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["id"],
                "properties": {
                  "id": { "type": "string" },
                  "name": { "type": "string" },
                  "email": { "type": "string" },
                  "phone": { "type": "string" },
                  "address": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": { "200": { "description": "Updated" } }
      },
      "delete": {
        "tags": ["Customers"],
        "summary": "Delete customer",
        "parameters": [{ "name": "id", "in": "query", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "Deleted" } }
      }
    },
    "/api/inventory": {
      "get": {
        "tags": ["Inventory"],
        "summary": "List inventory items",
        "responses": {
          "200": { "description": "Inventory list", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/InventoryItem" } } } } }
        }
      },
      "post": {
        "tags": ["Inventory"],
        "summary": "Create inventory item",
        "description": "Creates an inventory item with optional journal entry for purchase recording.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["sku", "name", "type", "unit", "quantity_on_hand", "cost_per_unit"],
                "properties": {
                  "sku": { "type": "string" },
                  "name": { "type": "string" },
                  "type": { "type": "string", "enum": ["raw", "finished"] },
                  "unit": { "type": "string" },
                  "quantity_on_hand": { "type": "number" },
                  "cost_per_unit": { "type": "number" },
                  "reorder_level": { "type": "number" },
                  "paymentSource": { "type": "string", "enum": ["cash", "bank", "payable", "opening"] }
                }
              }
            }
          }
        },
        "responses": { "201": { "description": "Created" } }
      },
      "put": {
        "tags": ["Inventory"],
        "summary": "Update inventory item",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["id"],
                "properties": {
                  "id": { "type": "string" },
                  "name": { "type": "string" },
                  "cost_per_unit": { "type": "number" },
                  "reorder_level": { "type": "number" }
                }
              }
            }
          }
        },
        "responses": { "200": { "description": "Updated" } }
      },
      "delete": {
        "tags": ["Inventory"],
        "summary": "Delete inventory item",
        "parameters": [{ "name": "id", "in": "query", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "Deleted" } }
      }
    },
...
    "/api/sales-orders": {
      "get": {
        "tags": ["Sales Orders"],
        "summary": "List sales orders",
        "parameters": [
          { "name": "limit", "in": "query", "schema": { "type": "number", "default": 50 } },
          { "name": "cursor", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": { "200": { "description": "Sales order list", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Pagination" } } } } }
      },
      "post": {
        "tags": ["Sales Orders"],
        "summary": "Create manual sales order",
        "description": "Creates a sales order with accounting journal entry.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["items", "customer_name", "total"],
                "properties": {
                  "items": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "product_id": { "type": "string" },
                        "product_name": { "type": "string" },
                        "quantity": { "type": "number" },
                        "unit_price": { "type": "number" },
                        "total": { "type": "number" }
                      }
                    }
                  },
                  "customer_name": { "type": "string" },
                  "total": { "type": "number" },
                  "notes": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": { "201": { "description": "Created" } }
      },
      "put": {
        "tags": ["Sales Orders"],
        "summary": "Update order status",
        "description": "If status='producing', auto-creates work order with design costs.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["orderId", "status"],
                "properties": {
                  "orderId": { "type": "string" },
                  "status": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": { "200": { "description": "Updated" } }
      }
    },
...
    "/api/invoices": {
      "get": {
        "tags": ["Invoices"],
        "summary": "List invoices",
        "parameters": [
          { "name": "limit", "in": "query", "schema": { "type": "number", "default": 50 } },
          { "name": "cursor", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": { "200": { "description": "Invoice list", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Pagination" } } } } }
      },
      "post": {
        "tags": ["Invoices"],
        "summary": "Create invoice",
        "description": "Creates invoice with revenue and COGS journal entries (double-entry).",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["amount", "total_amount", "customer_id", "customer_name"],
                "properties": {
                  "amount": { "type": "number", "description": "Subtotal before tax" },
                  "tax_amount": { "type": "number" },
                  "total_amount": { "type": "number", "description": "Total including tax" },
                  "cost_of_goods_sold": { "type": "number" },
                  "customer_id": { "type": "string" },
                  "customer_name": { "type": "string" },
                  "items": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "description": { "type": "string" },
                        "quantity": { "type": "number" },
                        "unit_price": { "type": "number" },
                        "total": { "type": "number" }
                      }
                    }
                  },
                  "due_date": { "type": "string", "format": "date" },
                  "sales_order_id": { "type": "string" },
                  "notes": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": { "201": { "description": "Created" } }
      }
    },
    "/api/payments": {
      "get": {
        "tags": ["Payments"],
        "summary": "List payments",
        "parameters": [
          { "name": "limit", "in": "query", "schema": { "type": "number", "default": 50 } },
          { "name": "cursor", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": { "200": { "description": "Payment list" } }
      },
      "post": {
        "tags": ["Payments"],
        "summary": "Record payment",
        "description": "Records a payment against an invoice and creates a journal entry within a Firestore transaction.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["amount", "invoice_id", "payment_method"],
                "properties": {
                  "amount": { "type": "number" },
                  "invoice_id": { "type": "string" },
                  "payment_method": { "type": "string", "enum": ["cash", "bank", "card"] },
                  "date": { "type": "string", "format": "date" },
                  "reference_number": { "type": "string" },
                  "notes": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": { "201": { "description": "Payment recorded" } }
      }
    },
    "/api/work-orders": {
      "get": {
        "tags": ["Work Orders"],
        "summary": "List all work orders with designs",
        "responses": { "200": { "description": "Work order list" } }
      },
      "post": {
        "tags": ["Work Orders"],
        "summary": "Create work order",
        "description": "Auto-calculates costs from design or from order items.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["quantity"],
                "properties": {
                  "salesOrderId": { "type": "string" },
                  "designId": { "type": "string" },
                  "quantity": { "type": "number" },
                  "assigned_worker": { "type": "string" },
                  "notes": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": { "201": { "description": "Created" } }
      },
      "put": {
        "tags": ["Work Orders"],
        "summary": "Update work order",
        "description": "Whitelisted fields only.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["id"],
                "properties": {
                  "id": { "type": "string" },
                  "status": { "type": "string" },
                  "completionPercentage": { "type": "number" },
                  "notes": { "type": "string" },
                  "assigned_worker": { "type": "string" },
                  "started_at": { "type": "string", "format": "date-time" },
                  "completed_at": { "type": "string", "format": "date-time" }
                }
              }
            }
          }
        },
        "responses": { "200": { "description": "Updated" } }
      }
    },
...
    "/api/designs": {
      "get": {
        "tags": ["Designs"],
        "summary": "List designs with filters",
        "parameters": [
          { "name": "category", "in": "query", "schema": { "type": "string" } },
          { "name": "status", "in": "query", "schema": { "type": "string" } },
          { "name": "complexity", "in": "query", "schema": { "type": "string" } },
          { "name": "minCost", "in": "query", "schema": { "type": "number" } },
          { "name": "maxCost", "in": "query", "schema": { "type": "number" } },
          { "name": "minMargin", "in": "query", "schema": { "type": "number" } },
          { "name": "pageSize", "in": "query", "schema": { "type": "number" } },
          { "name": "lastDocId", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": { "200": { "description": "Design list" } }
      },
      "post": {
        "tags": ["Designs"],
        "summary": "Create design",
        "responses": { "201": { "description": "Created" } }
      }
    },
...
    "/api/bom": {
      "get": {
        "tags": ["BOM"],
        "summary": "List BOMs",
        "parameters": [
          { "name": "designId", "in": "query", "schema": { "type": "string" } },
          { "name": "status", "in": "query", "schema": { "type": "string", "enum": ["draft", "active", "archived"] } },
          { "name": "limit", "in": "query", "schema": { "type": "number" } }
        ],
        "responses": { "200": { "description": "BOM list" } }
      },
      "post": {
        "tags": ["BOM"],
        "summary": "Create BOM",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["design_id", "name", "items"],
                "properties": {
                  "design_id": { "type": "string" },
                  "name": { "type": "string" },
                  "items": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "item_id": { "type": "string" },
                        "quantity": { "type": "number" },
                        "unit_cost": { "type": "number" }
                      }
                    }
                  },
                  "labor_hours": { "type": "number" },
                  "labor_rate": { "type": "number" },
                  "overhead_percentage": { "type": "number" },
                  "notes": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": { "201": { "description": "Created" } }
      }
    },
...
    "/api/vendors": {
      "get": {
        "tags": ["Vendors"],
        "summary": "List vendors",
        "parameters": [
          { "name": "status", "in": "query", "schema": { "type": "string", "enum": ["active", "inactive"] } },
          { "name": "search", "in": "query", "schema": { "type": "string" } },
          { "name": "minRating", "in": "query", "schema": { "type": "number" } }
        ],
        "responses": { "200": { "description": "Vendor list" } }
      },
      "post": {
        "tags": ["Vendors"],
        "summary": "Create vendor",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name"],
                "properties": {
                  "name": { "type": "string" },
                  "contact_name": { "type": "string" },
                  "email": { "type": "string" },
                  "phone": { "type": "string" },
                  "address": { "type": "string" },
                  "payment_terms": { "type": "string" },
                  "lead_time_days": { "type": "number" },
                  "notes": { "type": "string" },
                  "status": { "type": "string", "enum": ["active", "inactive"] }
                }
              }
            }
          }
        },
        "responses": { "201": { "description": "Created" } }
      }
    },
...
    "/api/purchase-orders": {
      "get": {
        "tags": ["Purchase Orders"],
        "summary": "List purchase orders",
        "parameters": [
          { "name": "vendorId", "in": "query", "schema": { "type": "string" } },
          { "name": "status", "in": "query", "schema": { "type": "string" } },
          { "name": "limit", "in": "query", "schema": { "type": "number" } }
        ],
        "responses": { "200": { "description": "PO list" } }
      },
      "post": {
        "tags": ["Purchase Orders"],
        "summary": "Create purchase order",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["vendor_id", "items"],
                "properties": {
                  "vendor_id": { "type": "string" },
                  "items": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "item_id": { "type": "string" },
                        "item_name": { "type": "string" },
                        "quantity": { "type": "number" },
                        "unit_cost": { "type": "number" }
                      }
                    }
                  },
                  "expected_delivery": { "type": "string", "format": "date" },
                  "shipping_address": { "type": "string" },
                  "shipping_cost": { "type": "number" },
                  "tax_rate": { "type": "number" },
                  "notes": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": { "201": { "description": "Created" } }
      }
    },
...
    "/api/chart-of-accounts": {
      "get": {
        "tags": ["Chart of Accounts"],
        "summary": "Get all accounts with live balances",
        "description": "Returns all accounts with balances derived live from journal entries.",
        "responses": {
          "200": {
            "description": "Chart of accounts",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "accounts": { "type": "array", "items": { "type": "object" } },
                    "journalEntries": { "type": "array", "items": { "type": "object" } }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Chart of Accounts"],
        "summary": "Add custom account",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["type", "data"],
                "properties": {
                  "type": { "type": "string", "description": "Always 'account'" },
                  "data": {
                    "type": "object",
                    "properties": {
                      "code": { "type": "string" },
                      "name": { "type": "string" },
                      "type": { "type": "string" },
                      "subType": { "type": "string" },
                      "normalBalance": { "type": "string", "enum": ["debit", "credit"] }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": { "201": { "description": "Created" } }
      }
    },
...
    "/api/accounting/expenses": {
      "get": {
        "tags": ["Accounting"],
        "summary": "List recorded expenses",
        "responses": { "200": { "description": "Expense list" } }
      },
      "post": {
        "tags": ["Accounting"],
        "summary": "Record business expense",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["amount", "description", "expenseAccount", "paymentMethod"],
                "properties": {
                  "amount": { "type": "number" },
                  "description": { "type": "string" },
                  "expenseAccount": { "type": "string" },
                  "paymentMethod": { "type": "string", "enum": ["cash", "bank"] }
                }
              }
            }
          }
        },
        "responses": { "201": { "description": "Expense recorded" } }
      }
    },
...
    "/api/accounting/liabilities": {
      "get": {
        "tags": ["Accounting"],
        "summary": "List liabilities",
        "responses": { "200": { "description": "Liability list" } }
      },
      "post": {
        "tags": ["Accounting"],
        "summary": "Record liability (incur or repay)",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["amount", "description", "liabilityAccount", "transactionType"],
                "properties": {
                  "amount": { "type": "number" },
                  "description": { "type": "string" },
                  "liabilityAccount": { "type": "string" },
                  "offsetAccount": { "type": "string" },
                  "transactionType": { "type": "string", "enum": ["incur", "repay"] }
                }
              }
            }
          }
        },
        "responses": { "201": { "description": "Liability recorded" } }
      }
    },
...
  }
}